// apps/api/src/common/common.constant.ts
export const RADIX = 10;

export const DATE_FORMAT = 'YYYY-MM-DD';

export enum SwaggerTag {
  Setting = 'Setting',
  Auth = 'Auth',
  User = 'User',
  Booking = 'Booking',
}

export enum PaginationPage {
  Min = 1,
}

export enum PaginationSize {
  Min = 1,
  Max = 500,
  Default = 10,
}

export enum SortDirection {
  Asc = 'asc',
  Desc = 'desc',
}