// apps/api/src/booking/booking.constant.ts
export enum SurfingExperience {
  Min = 0,
  Max = 10,
}

export enum BookingSortProperty {
  Id = 'id',
  Name = 'name',
  Email = 'email',
  Phone = 'phone',
  CountryId = 'country_id',
  SurfingExperience = 'surfing_experience',
  Date = 'date',
  SurfboardId = 'surfboard_id',
  CreatedAt = 'created_at',
  UpdatedAt = 'updated_at',
  DeletedAt = 'deleted_at',
}